# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar-IndigoPay API
  version: 1.0.0
  description: |
    OpenAPI 3.0 documentation for the backend REST API used by Stellar-IndigoPay.
    The API exposes project discovery, donations, profiles, leaderboard, notifications,
    ratings, updates, jobs, stats, and admin management endpoints.
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Local development server
  - url: https://api.stellarindigopay.com
    description: Production server
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Authentication is required or the provided token is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid request payload
      required:
        - error
...
    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            token:
              type: string
            refreshToken:
              type: string
            expiresIn:
              type: integer
    AdminRefreshResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            token:
              type: string
            expiresIn:
              type: integer
    AdminMeResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            username:
              type: string
            role:
              type: string
    AdminAuditLogResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              actor:
                type: string
              action:
                type: string
              target_type:
                type: string
              target_id:
                type: string
              metadata:
                type: object
              ip_address:
                type: string
              created_at:
                type: string
                format: date-time
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
...
    AdminLoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AdminRefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
...
paths:
...
  /api/admin/login:
    post:
      tags: [Admin]
      summary: Log in as an administrator
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Admin authentication tokens.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/admin/refresh:
    post:
      tags: [Admin]
      summary: Refresh an admin token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRefreshRequest"
      responses:
        "200":
          description: Refreshed admin token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRefreshResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/admin/me:
    get:
      tags: [Admin]
      summary: Get current admin identity
      responses:
        "200":
          description: Current admin identity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/admin/audit-log:
    get:
      tags: [Admin]
      summary: Query admin audit log
      parameters:
        - name: actor
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Audit log page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuditLogResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...